Shared logic with normal TS deploy
