+++ src/api/deploy-test-suite/helpers/record-test-run.js
@@ -7,7 +7,9 @@ async function createRecordTestRun(imageName, runId, environment, user) {
 
   const url = `${config.get('portalBackendUrl')}/test-run`
 
-  logger.info(`Recording  test-run for ${imageName} run ${runId} by ${user}`)
+  logger.info(
+    `Recording  test-run for ${imageName} run ${runId} by ${user.displayName}`
+  )
   return await fetch(url, {
     method: 'POST',
     headers: {
